requirements.txt
Django==1.4.1
Pillow==1.7.8
argparse==1.2.1
distribute==0.6.28
django-registration==0.8
wsgiref==0.1.2
